Fix compiler warnings across codebase
- Fix sign conversion warnings (-Wsign-conversion) in multiple files:
  - SegmentedButtonBar, OscilDropdown, OscilTabs, PluginTestServer
  - OscillatorConfigPopup, Pane, SharedCaptureBuffer, SignalProcessor
  - WaveformComponent, PluginEditor
- Fix shadow variable warning in SourceCoordinator
- Fix switch-enum warning in TimingEngine
- Fix deprecated volatile usage in test_shared_capture_buffer
- Fix unused parameter/lambda capture warnings
- Migrate deprecated juce::Font(float) to FontOptions API in:
  - OscilTextField, OscilDropdown, OscilTabs, OscilButton
  - OscilToggle, OscilSlider, OscilCheckbox, OscilRadioButton
  - OscilTooltip
- Fix destructor override in OscilTooltip.h

All 165 tests pass.

+++ include/dsp/SignalProcessor.h
@@ -31,9 +31,9 @@ struct ProcessedSignal
     {
         numSamples = samples;
         isStereo = stereo;
-        channel1.resize(samples);
+        channel1.resize(static_cast<size_t>(samples));
         if (stereo)
-            channel2.resize(samples);
+            channel2.resize(static_cast<size_t>(samples));
         else
             channel2.clear();
     }
